Move Math related code to Math folder
