+++ NAS2D/Math/MathUtils.cpp
