+++ NAS2D/Math/Point.cpp
