+++ NAS2D/Math/PointInRectangleRange.h
