+++ NAS2D/Math/Rectangle.cpp
