+++ NAS2D/Math/Rectangle.h
