+++ NAS2D/Math/Trig.cpp
