+++ NAS2D/Math/VectorSizeRange.h
