+++ NAS2D/NAS2D.vcxproj
@@ -200,7 +200,7 @@
     <ClCompile Include="Filesystem.cpp" />
     <ClCompile Include="FpsCounter.cpp" />
     <ClCompile Include="Game.cpp" />
-    <ClCompile Include="MathUtils.cpp" />
+    <ClCompile Include="Math\MathUtils.cpp" />
     <ClCompile Include="ParserHelper.cpp" />
     <ClCompile Include="Mixer\Mixer.cpp" />
     <ClCompile Include="Mixer\MixerSDL.cpp" />
@@ -208,8 +208,8 @@
     <ClCompile Include="Renderer\Color.cpp" />
     <ClCompile Include="Renderer\DisplayDesc.cpp" />
     <ClCompile Include="Renderer\Fade.cpp" />
-    <ClCompile Include="Renderer\Point.cpp" />
-    <ClCompile Include="Renderer\Rectangle.cpp" />
+    <ClCompile Include="Math\Point.cpp" />
+    <ClCompile Include="Math\Rectangle.cpp" />
     <ClCompile Include="Renderer\RectangleSkin.cpp" />
     <ClCompile Include="Renderer\Renderer.cpp" />
     <ClCompile Include="Renderer\RendererOpenGL.cpp" />
@@ -222,7 +222,7 @@
     <ClCompile Include="StateManager.cpp" />
     <ClCompile Include="StringUtils.cpp" />
     <ClCompile Include="Timer.cpp" />
-    <ClCompile Include="Trig.cpp" />
+    <ClCompile Include="Math\Trig.cpp" />
     <ClCompile Include="Version.cpp" />
     <ClCompile Include="Xml\XmlNode.cpp" />
     <ClCompile Include="Xml\XmlAttribute.cpp" />
@@ -246,7 +246,7 @@
     <ClInclude Include="Filesystem.h" />
     <ClInclude Include="FpsCounter.h" />
     <ClInclude Include="Game.h" />
-    <ClInclude Include="MathUtils.h" />
+    <ClInclude Include="Math\MathUtils.h" />
     <ClInclude Include="Mixer\Mixer.h" />
     <ClInclude Include="Mixer\MixerSDL.h" />
     <ClInclude Include="Mixer\MixerNull.h" />
@@ -256,11 +256,11 @@
     <ClInclude Include="Renderer\RendererNull.h" />
     <ClInclude Include="Renderer\Color.h" />
     <ClInclude Include="Renderer\Fade.h" />
-    <ClInclude Include="Renderer\Point.h" />
-    <ClInclude Include="Renderer\PointInRectangleRange.h" />
-    <ClInclude Include="Renderer\Vector.h" />
-    <ClInclude Include="Renderer\VectorSizeRange.h" />
-    <ClInclude Include="Renderer\Rectangle.h" />
+    <ClInclude Include="Math\Point.h" />
+    <ClInclude Include="Math\PointInRectangleRange.h" />
+    <ClInclude Include="Math\Vector.h" />
+    <ClInclude Include="Math\VectorSizeRange.h" />
+    <ClInclude Include="Math\Rectangle.h" />
     <ClInclude Include="Renderer\RectangleSkin.h" />
     <ClInclude Include="Renderer\Renderer.h" />
     <ClInclude Include="Renderer\RendererOpenGL.h" />
@@ -280,7 +280,7 @@
     <ClInclude Include="StringUtils.h" />
     <ClInclude Include="StringValue.h" />
     <ClInclude Include="Timer.h" />
-    <ClInclude Include="Trig.h" />
+    <ClInclude Include="Math\Trig.h" />
     <ClInclude Include="Utility.h" />
     <ClInclude Include="Version.h" />
     <ClInclude Include="Xml\Xml.h" />
