+++ NAS2D/Renderer/Fade.cpp
@@ -11,7 +11,7 @@
 
 #include "Fade.h"
 #include "Renderer.h"
-#include "Rectangle.h"
+#include "../Math/Rectangle.h"
 
 #include <algorithm>
