+++ NAS2D/Renderer/Renderer.h
@@ -12,8 +12,8 @@
 
 #include "Color.h"
 #include "DisplayDesc.h"
-#include "Point.h"
-#include "Vector.h"
+#include "../Math/Point.h"
+#include "../Math/Vector.h"
 #include "../Timer.h"
 #include "../Signal/Signal.h"
