+++ NAS2D/Resource/Image.cpp
@@ -9,7 +9,7 @@
 // ==================================================================================
 #include "Image.h"
 
-#include "../Renderer/Rectangle.h"
+#include "../Math/Rectangle.h"
 #include "../Filesystem.h"
 #include "../Utility.h"
