+++ test-graphics/TestGraphics.cpp
@@ -2,7 +2,7 @@
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
-#include <NAS2D/Renderer/Rectangle.h>
+#include <NAS2D/Math/Rectangle.h>
 
 #include <functional>
 #include <random>
