+++ test/Math/Point.test.cpp
@@ -1,4 +1,4 @@
-#include "NAS2D/Renderer/Point.h"
+#include "NAS2D/Math/Point.h"
 #include <gtest/gtest.h>
