+++ test/Math/Rectangle.test.cpp
@@ -1,4 +1,4 @@
-#include "NAS2D/Renderer/Rectangle.h"
+#include "NAS2D/Math/Rectangle.h"
 #include <gtest/gtest.h>
