+++ test/test.vcxproj
@@ -46,16 +46,16 @@
   </PropertyGroup>
   <ItemGroup>
     <ClCompile Include="Renderer/Color.test.cpp" />
-    <ClCompile Include="Renderer/Point.test.cpp" />
-    <ClCompile Include="Renderer/PointInRectangleRange.test.cpp" />
-    <ClCompile Include="Renderer/Rectangle.test.cpp" />
-    <ClCompile Include="Renderer/Vector.test.cpp" />
-    <ClCompile Include="Renderer/VectorSizeRange.test.cpp" />
+    <ClCompile Include="Math/Point.test.cpp" />
+    <ClCompile Include="Math/PointInRectangleRange.test.cpp" />
+    <ClCompile Include="Math/Rectangle.test.cpp" />
+    <ClCompile Include="Math/Vector.test.cpp" />
+    <ClCompile Include="Math/VectorSizeRange.test.cpp" />
     <ClCompile Include="Resource/Image.test.cpp" />
     <ClCompile Include="Resource/ResourceCache.test.cpp" />
     <ClCompile Include="ContainerUtils.test.cpp" />
     <ClCompile Include="Filesystem.test.cpp" />
-    <ClCompile Include="MathUtils.test.cpp" />
+    <ClCompile Include="Math/MathUtils.test.cpp" />
     <ClCompile Include="Signal/SignalConnection.test.cpp" />
     <ClCompile Include="Signal/Delegate.test.cpp" />
     <ClCompile Include="Signal/Signal.test.cpp" />
